requirements.txt
python-dotenv>=0.19.0
requests>=2.26.0
pandas>=1.3.0
openpyxl>=3.0.7
python-magic>=0.4.24
pathlib
zipfile36
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
langchain-openai
langchain-core
openai
tiktoken

# FastAPI dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-multipart>=0.0.6
email-validator>=2.0.0

# Database and storage
supabase>=2.0.0
psycopg2-binary>=2.9.7

# Utility libraries
httpx>=0.24.1

# Auth0 integration
auth0-python>=4.4.0
python-jose[cryptography]>=3.3.0
authlib>=1.2.1
starlette-csrf>=1.4.0

# Security libraries
cryptography>=40.0.0
fastapi-limiter>=0.1.5
bandit>=1.7.5
sqlalchemy>=2.0.0
pyOpenSSL>=23.0.0
aioredis>=2.0.0